% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ambulance_boundary_data}
\alias{ambulance_boundary_data}
\title{English ambulance services' boundaries data.}
\format{A spatial polygons data frame with 11 rows and 1 data variables:
\describe{
  \item{service}{ambulance service, short code (e.g. EMAS)}
}}
\usage{
ambulance_boundary_data
}
\description{
A dataset containing the boundaries of English ambulance services derived
from:
\itemize{
  \item The ONS / OS's 2011 LSOA boundaries
(\url{http://geoportal.statistics.gov.uk/datasets/da831f80764346889837c72508f046fa_2});
  \item the ONS's 2011 LSOA to 2016 CCG lookup
(\url{http://ons.maps.arcgis.com/home/item.html?id=19e5c35c6a504a7b9e1b74bed1b6225f}); and
  \item NHS England's 2016 CCG to ambulance service lookup
(\url{https://www.england.nhs.uk/statistics/wp-content/uploads/sites/2/2013/04/AmbSYS-TimeSeries-Interactive-Web-File-September-2016.xlsx}).
}
}
\details{
Contains National Statistics data: Crown copyright and database right 2016;
Contains OS data: Crown copyright and database right 2016.
}
\keyword{datasets}

