% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_boundary_data}
\alias{county_boundary_data}
\title{English counties' (UTLA) boundaries data.}
\format{A spatial polygons data frame with 150 rows and 3 data variables:
\describe{
  \item{CNTY14}{ONS County/UTLA code}
  \item{county}{County/UTLA name}
  \item{in_london}{Boolean, is the county/UTLA a London Borough?}
}}
\usage{
county_boundary_data
}
\description{
A dataset containing the boundaries of English counties (Upper Tier Local
Authorities) derived from:
\itemize{
  \item The ONS / OS's 2011 LSOA boundaries
(\url{http://geoportal.statistics.gov.uk/datasets/da831f80764346889837c72508f046fa_2});
  \item 2011 LSOA to 2011 Local Authority District lookup from
\url{http://geoconvert.mimas.ac.uk} [retrieved 27/09/2016];
  \item the ONS's 2014 Non-Metropolitan District to 2014 County lookup;
  \item the ONS's 2014 Metropolitan District names;
  \item the ONS's 2014 London Borough lookup names; and
  \item the ONS's 2014 Unitary Authority names.
 }
 Items above with unspecified source from ONS's Code History Database
\url{http://ons.maps.arcgis.com/home/item.html?id=cb98d1fdaa3c462a94e61de1a7eeaf8b}.
}
\details{
Minor alterations to 2011 LADs relating to:
\itemize{
  \item The St Albans and Welwyn Hatfield (Boundary Change) Order 2012;
  \item The East Hertfordshire and Stevenage (Boundary Change) Order 2013;
  \item The Gateshead and Northumberland (Boundary Change) Order 2013.
 }

Within this data:
\itemize{
  \item the London Boroughss of City of London and Westminster have been merged; and
  \item the Unitary Authorities of Cornwall and The Isles of Scilly have been merged.
 }

Contains National Statistics data: Crown copyright and database right 2016;
Contains OS data: Crown copyright and database right 2016.
}
\keyword{datasets}

