% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth-map-fns.R
\name{saveMaps}
\alias{saveMaps}
\title{Save choropleth map(s)}
\usage{
saveMaps(fname, fwidth, data, variable, title, qtiles = 5, breaks = NA,
  greyscale = FALSE)
}
\arguments{
\item{fname}{file name to save the plots under (no directory paths or file
extensions)}

\item{fwidth}{the width (pixels) of the plots to be produced}

\item{data}{A spatial polygons data frame. The map produced will be at the
county level if the column name \code{county} is present otherwise the map
will be at the level of the provided geography.}

\item{variable}{the name of the variable on the basis of which the map
should be themed}

\item{title}{the title to display on the map}

\item{qtiles}{the number of quantiles to split \code{variable} into. Note:
if \code{breaks} is specified, \code{qtiles} is ignored. By default
\code{variable} is split into quintiles.}

\item{breaks}{the (ordered) breaks on which to split \code{variable}, for
\code{n} splits \code{n+1} breaks must be specified. If \code{breaks} is
specified (other than \code{NA}), \code{qtiles} are ignored.}

\item{greyscale}{boolean; should the plot be in greyscale only? Default is
\code{FALSE}.}
}
\value{
\code{Save complete.} upon completion.  The saved files will appear
in the working directory.
}
\description{
This function will attempt to save choropleth map(s) using the data provided
to it. If the data is at the County level it will save both England and
London-only maps.
}
\details{
If \code{breaks} is specified, \code{qtiles} is ignored.
There is a maximum of:
\itemize{
 \item 9 levels allowed for greyscale plots;
 \item 11 levels for colour plots.
}
}
\examples{
\dontrun{
# example county-level data
county_data <- data.frame(county = county_boundary_data$county,
                          deaths = round(rnorm(length(county_boundary_data$county), 30, 8)))
# example ambulance-level data
service_data <- data.frame(service = ambulance_boundary_data$service,
                           hoaxes = round(rnorm(length(ambulance_boundary_data$service), 70, 8)))
service_data <- service_data[service_data$service != "IOW", ]

# merge numeric data and spatial data
county_data_spatial <- merge(county_boundary_data, county_data, by = "county")
service_data_spatial <- merge(ambulance_boundary_data, service_data, by = "service")

# save the nice maps
saveMaps("lovely-map", 500, county_data, "deaths", "My lovely map")
}
}

