% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanmaps.R
\docType{package}
\name{vanmaps}
\alias{vanmaps}
\alias{vanmaps-package}
\title{vanmaps: A package for producing choropleth maps at the English County
or Ambulance Service level.}
\description{
Contains National Statistics data: Crown copyright and database right 2016;
Contains OS data: Crown copyright and database right 2016
}

